+++ main.go
@@ -237,11 +237,29 @@ func main() {
 			return err
 		}
 		for _, set := range args {
-			k, v := keyPrompt(set)
+			k, v := keyPrompt(set, true)
 			s.Set(k, v)
 		}
 		return v.Write(path, s)
 	}, "write")
+
+	r.Dispatch("paste", func(command string, args ...string) error {
+		rc.Apply()
+		if len(args) < 2 {
+			return fmt.Errorf("USAGE: set path key[=value] [key ...]")
+		}
+		v := connect()
+		path, args := args[0], args[1:]
+		s, err := v.Read(path)
+		if err != nil && err != vault.NotFound {
+			return err
+		}
+		for _, set := range args {
+			k, v := keyPrompt(set, false)
+			s.Set(k, v)
+		}
+		return v.Write(path, s)
+	})
 
 	r.Dispatch("get", func(command string, args ...string) error {
 		rc.Apply()