+++ ci/release_notes.md
@@ -6,3 +6,6 @@
   against the Vault (using either token, Github or LDAP backends)
 - The new `safe env` sub-command exists to help troubleshoot
   auth/targeting issues.
+- New `safe paste` sub-command is just like `safe set`, except it
+  never asks for confirmation of what you just pasted from
+  1password.